Added an Environment child for the recipe
environment described in xcookybooky. That looks
like it works when we graft it into the PyLaTeX
document like this. Called out xcookybooky and
PyLaTeX in the REAMDE, updated requirements.txt
with PyLaTeX. Added an env target to Makefile for
the standard non-dev environment. Also added a
clean target for taking care of the TeX build
artifacts in case we have to hit the .tex that
PyLaTeX comes up with pdflatex manually to flush
the bogs.