Simplified README description,
Attempted to integrate the support command \step
from the \preparation environment but that got all
kinds of mad. Opted to swap to Enumerate for that.
Tweaked Ropa Vieja recipe and added two more.
Integrated reading the json directory into the
vittles class and am starting to populate the
document based on the json files.