Added a category subfolder to json
that can then be used to create Sections.
Added a requirements directory to the project
root and moved the venv definitions into there.
Added dorito casserole to recipes. Tweaked the
recipe adder to accept categories. Added fields
to required fields for recipe, as well as
implementing more attributes to correspond
to those fields. Added Portion command as well
as cookingsymbols package.