Changes centered around making
the images less cumbersome to deal with.
Added an ImageParamAdder that'll dump
a json file which can be tweaked to make
the images show up nicely on their given
pages. Scaling, rotation and following vspace
seem to be enough to get all of the images
I've tested with so far to look nice. Went
with scripts in img that link to subdirs to
hook up to the actual images per recipe, this
allows, for example, personalized recipe
books to be generated for different sets of
relatives.